Refactor codebase to a modular structure and implement database for usage tracking
