+++ telegram/bot.py
@@ -1,0 +1,4 @@
+from telebot.async_telebot import AsyncTeleBot
+from config import TELEGRAM_TOKEN
+
+bot = AsyncTeleBot(TELEGRAM_TOKEN)
