+++ twitter/utils.py
@@ -1,4 +1,3 @@
-import tweepy
 import aiohttp
 import logging
 import time
@@ -7,6 +6,8 @@ import os
 import asyncio
 from functools import wraps
 from config import MONTHLY_WRITE_CAP, USAGE_FILE
+from twitter.api import client, api
+import tweepy
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
@@ -32,7 +33,7 @@ def retry_on_rate_limit(max_retries=3, backoff_factor=2):
     return decorator
 
 @retry_on_rate_limit()
-async def post_tweet_with_media(client, api, text, media_paths=None, reply_to_id=None):
+async def post_tweet_with_media(text, media_paths=None, reply_to_id=None):
     media_ids = []
     if media_paths:
         for path in media_paths:
@@ -56,29 +57,46 @@ async def post_tweet_with_media(client, api, text, media_paths=None, reply_to_id
         logging.error(f"Post failed: {e}")
         raise
 
+import sqlite3
+
+def init_db():
+    conn = sqlite3.connect('usage.db')
+    c = conn.cursor()
+    c.execute('''
+        CREATE TABLE IF NOT EXISTS usage (
+            month TEXT PRIMARY KEY,
+            writes INTEGER
+        )
+    ''')
+    conn.commit()
+    conn.close()
+
 def track_usage(reset=False):
-    if reset or not os.path.exists(USAGE_FILE):
-        data = {"writes": 0, "month": time.strftime("%Y-%m")}
-    else:
-        try:
-            with open(USAGE_FILE, "r") as f:
-                data = json.load(f)
-        except:
-            data = {"writes": 0, "month": time.strftime("%Y-%m")}
+    init_db()
+    conn = sqlite3.connect('usage.db')
+    c = conn.cursor()
 
     current_month = time.strftime("%Y-%m")
-    if data["month"] != current_month:
-        data = {"writes": 0, "month": current_month}
+
+    c.execute("SELECT writes FROM usage WHERE month = ?", (current_month,))
+    result = c.fetchone()
+
+    if result is None:
+        c.execute("INSERT INTO usage (month, writes) VALUES (?, ?)", (current_month, 0))
+        writes = 0
+    else:
+        writes = result[0]
 
     if not reset:
-        data["writes"] += 1
-        if data["writes"] > MONTHLY_WRITE_CAP * 0.9:
-            logging.warning(f"Near cap: {data['writes']}/{MONTHLY_WRITE_CAP}")
+        writes += 1
+        if writes > MONTHLY_WRITE_CAP * 0.9:
+            logging.warning(f"Near cap: {writes}/{MONTHLY_WRITE_CAP}")
 
-    with open(USAGE_FILE, "w") as f:
-        json.dump(data, f)
+    c.execute("UPDATE usage SET writes = ? WHERE month = ?", (writes, current_month))
+    conn.commit()
+    conn.close()
 
-    return data["writes"]
+    return writes
 
 async def download_media(url):
     async with aiohttp.ClientSession() as session:
