--- usage.json
@@ -1,1 +1,0 @@
-{"writes": 4, "month": "2025-07"}
